refactor: modularize App into smaller components
